+++ ui/packages/shared/parser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.77 (2024-07-01)
+
+**Note:** Version bump only for package @parca/parser
+
+## 0.16.76 (2024-06-27)
+
+**Note:** Version bump only for package @parca/parser
+
 ## 0.16.75 (2024-06-06)
 
 **Note:** Version bump only for package @parca/parser