+++ ui/packages/shared/profile/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.398](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.397...@parca/profile@0.16.398) (2024-07-01)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.397](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.396...@parca/profile@0.16.397) (2024-07-01)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.396](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.395...@parca/profile@0.16.396) (2024-07-01)
+
+**Note:** Version bump only for package @parca/profile
+
+## 0.16.395 (2024-07-01)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.394](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.393...@parca/profile@0.16.394) (2024-06-27)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.393](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.392...@parca/profile@0.16.393) (2024-06-27)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.392](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.391...@parca/profile@0.16.392) (2024-06-27)
+
+**Note:** Version bump only for package @parca/profile
+
+## 0.16.391 (2024-06-24)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.390](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.389...@parca/profile@0.16.390) (2024-06-20)
+
+**Note:** Version bump only for package @parca/profile
+
 ## 0.16.389 (2024-06-18)
 
 **Note:** Version bump only for package @parca/profile